pyproject.toml
[build-system]
requires = ["hatchling>=1.26.0"]
build-backend = "hatchling.build"

[project]
name = "kdrifting"
version = "0.1.0"
description = "A faithful, high-quality PyTorch port of the Drift research release."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Ayan Das" }]
dependencies = [
  "einops>=0.8.0",
  "huggingface-hub>=0.25.0",
  "numpy>=1.26,<3",
  "pillow>=10.4.0",
  "pyyaml>=6.0.2",
  "safetensors>=0.4.5",
  "scipy>=1.13.1",
  "torch>=2.4.0",
  "torchvision>=0.19.0",
  "tqdm>=4.66.0",
  "wandb>=0.18.0",
]

[project.optional-dependencies]
dev = [
  "flax>=0.10.4",
  "jax[cpu]>=0.4.36",
  "optax>=0.2.5",
  "orbax-checkpoint>=0.11.0",
  "pre-commit>=4.0.1",
  "pyright>=1.1.389",
  "pytest>=8.3.3",
  "pytest-cov>=5.0.0",
  "ruff>=0.8.0",
]
hf = [
  "diffusers>=0.30.0",
  "transformers>=4.45.0",
]
jax = [
  "flax>=0.10.4",
  "jax[cpu]>=0.4.36",
  "optax>=0.2.5",
  "orbax-checkpoint>=0.11.0",
]

[project.scripts]
kdrifting = "kdrifting.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["kdrifting"]

[tool.pytest.ini_options]
addopts = [
  "--strict-config",
  "--strict-markers",
  "-ra",
]
filterwarnings = ["error"]
pythonpath = ["."]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.format]
docstring-code-format = true

[tool.ruff.lint]
select = [
  "B",
  "E",
  "F",
  "I",
  "N",
  "PERF",
  "PIE",
  "PTH",
  "RUF",
  "SIM",
  "UP",
  "W",
]
ignore = [
  "B008",
]

[tool.ruff.lint.isort]
known-first-party = ["kdrifting"]

[tool.pyright]
include = ["kdrifting", "tests"]
exclude = [".git", ".pytest_cache", ".ruff_cache", ".venv"]
venvPath = "."
venv = ".venv"
extraPaths = [
  ".venv/lib/python3.11/site-packages",
  "/run/current-system/profile/lib/python3.11/site-packages",
]
pythonVersion = "3.11"
typeCheckingMode = "strict"
reportMissingTypeStubs = "none"
reportUnknownVariableType = "error"
reportUnknownMemberType = "error"
reportUnknownArgumentType = "error"
reportUnknownLambdaType = "error"
reportUnknownParameterType = "error"
reportMissingParameterType = "error"
reportMissingTypeArgument = "error"
reportImportCycles = "error"
